pyproject.toml
[build-system]
requires = [
  "setuptools>=82",
  "setuptools_scm>=8"
]
build-backend = "setuptools.build_meta"

[project]
name = "sendnn-inference"
description = "vLLM plugin for Spyre hardware support"
readme = "README.md"
license = {text = "Apache 2"}
dependencies = [
    "fms-model-optimizer[fp8]>=0.8.2,<0.9",
    "ibm-fms>=1.9.0,<2",
    # NB: use strict < with the next patch version to not exclude versions with
    # build metadata suffixes
    "vllm>=0.19.0,<0.19.2",
    
    # Specific torch version overrides handled by uv
    "torch",
    "torchvision",
]
requires-python = ">=3.11"
dynamic = ["version"]

[project.entry-points."vllm.platform_plugins"]
sendnn_inference = "sendnn_inference:register"

[tool.setuptools.packages.find]
where = ["."]  # list of folders that contain the packages (["."] by default)
include = ["sendnn_inference*"]  # package names should match these glob patterns (["*"] by default)
exclude = []  # exclude packages matching these glob patterns (empty by default)
namespaces = false  # to disable scanning PEP 420 namespaces (true by default)

[tool.setuptools_scm]
# Version files are optional, this just marginally speeds up version inspection
# at runtime
version_file = "sendnn_inference/_version.py"
# Version strings are formatted as `{version}+{local}`, e.g. "0.1.dev1+gb1a7032"
# Local versioning is incompatible with pypi, so we disable it by default.
local_scheme = "no-local-version"
# Only match vX.Y.Z tags (with optional SemVer pre-release suffix)
# SemVer pre-release format: hyphen followed by dot-separated identifiers (alphanumeric + hyphens)
tag_regex = "^v(?P<version>[0-9]+\\.[0-9]+\\.[0-9]+(?:-[0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*)?)$"
git_describe_command = "git describe --dirty --tags --long --match 'v*'"
# For more version configuration, see
# https://setuptools-scm.readthedocs.io/en/latest/config/


[tool.uv]
# Triton is optional in most dependencies, however, xgrammar requires it.
# So we set to 'never' to prevent installing it or eventually any other
# by accident
override-dependencies = [
    "torch==2.10.0",
    "triton; sys_platform == 'never'",
    "intel-extension-for-pytorch; sys_platform == 'never'",

    # opencv-python-headless==4.13.0.90 has a packaging bug that incorrectly depends on libxcb.so.1,
    # causing import failure on headless Linux systems.
    "opencv-python-headless==4.12.0.88",

    # Skip packages on s390x and ppc64le, expected to be pre-installed
    "vllm ; platform_machine not in 's390x, ppc64le'",
    "ray; platform_machine not in 's390x, ppc64le'",
    "pyarrow; platform_machine not in 's390x, ppc64le'",

    # uv doesn't resolve numba->llvmlite constraint correctly...
    #   if numba is updated, this will likely need to be too
    "llvmlite==0.44.0; platform_machine not in 's390x, ppc64le'",

    # setuptools is specified as a runtime dep by vllm, make sure it's patched for CVEs
    "setuptools>=82",

    # requests is used by many dependencies, make sure it's patched for CVEs
    "requests>=2.32.4",

    # vllm 0.18.0 pins compressed-tensors==0.13.0; override to 0.14.0.1
    # TODO: remove once minimum vllm is bumped past 0.18.0
    "compressed-tensors==0.14.0.1",

    # llguidance>=1.7.3 fixes s390x endianness issues.
    # This conflicts with vLLM's version range (llguidance >= 1.3.0, < 1.4.0).
    # TODO: Remove this override once vLLM's requirement range includes or moves past 1.7.3.
    # See: https://github.com/vllm-project/vllm/blob/v0.19.1/requirements/common.txt#L22
    "llguidance>=1.7.3"
]
# This adds constraints to all dependent build environments, which will ensure everything is built
# with the same version of torch. This CANNOT conflict with a package's existing build dependencies
build-constraint-dependencies = ["torch==2.10.0"]
# This ensures that we build with the empty backend for vLLM
extra-build-variables = { vllm = { VLLM_TARGET_DEVICE = "empty" } }

[tool.uv.sources]
vllm = { git = "https://github.com/vllm-project/vllm", rev = "v0.19.1" }
torch = [
  { index = "pytorch-cpu" },
]
torchvision = [
  { index = "pytorch-cpu" },
]

# The pytorch-cpu index here ensures we always pull the cpu version of torch everywhere
# The `explicit=true` setting means this index will only be used for packages that we explicitly set
# it for in `tool.uv.sources`. This is important to avoid requiring the use of 
# `--index-strategy unsafe-best-match`
[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ty.rules]
possibly-missing-attribute = "ignore"

[tool.ty.src]
exclude = ["sendnn_inference/__init__.py"]

[tool.ruff]
# Use widescreen monitors 😎😎😎
line-length = 100
exclude = [
    "sendnn_inference/model_executor/model_loader/spyre_setup.py"
]

[tool.ruff.lint.per-file-ignores]
"sendnn_inference/version.py" = ["F401"]
"sendnn_inference/_version.py" = ["ALL"]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    # "I",
    "G",
]
ignore = [
    # star imports
    "F405", "F403",
    # lambda expression assignment
    "E731",
    # Loop control variable not used within loop body
    "B007",
    # f-string format
    "UP032",
    # % formatters
    "UP031",
    # ternary operators
    "SIM108",


    # TODO: Fix all of these in code instead

    # Type unions
    "UP007",
    # Callable
    "UP035",
    # Zip
    "B905",
]

[tool.codespell]
ignore-words-list = "dout, te, indicies, subtile, ElementE"
skip = "*.svg,./tests/hf_cache.json"
#skip = "./tests/models/fixtures,./tests/prompts,./benchmarks/sonnet.txt,./tests/lora/data,./build"

[tool.isort]
use_parentheses = true
skip_gitignore = true

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_default_fixture_loop_scope = "function"

# --8<-- [start:test-markers-definition]
markers = [
    "skip_global_cleanup",

    # Test categories
    "e2e: Tests using end-to-end engine spin-up",
    "utils: Tests for utility functions",
    "worker: Tests for worker logic",
    "basic: Basic correctness tests",
    "compat: Tests readiness to remove backwards compatibility code",
    "precompilation: Tests related to precompiled model cache",

    # Hardware backends
    "cpu: Tests using CPU (i.e. eager) backend",
    "spyre: Tests using Spyre hardware backend",

    # Operational modes
    "cb: Continuous batching tests",
    "chunked_prefill: Tests with chunked prefill enabled",
    "prefix_caching: Tests for prefix caching",
    "multi: Tests that require >1 cards",

    # Model types
    "decoder: Tests for decoder models",
    "embedding: Tests for embedding models",
    "quantized: Tests for quantized models",
    "full_model: Tests with a full-sized decoder model (instead of a micro model)",
    "scoring: Tests for scoring models",
    "multimodal: Tests for multimodal models",

]
# --8<-- [end:test-markers-definition]

[tool.markdownlint]
#MD013.line_length = 100
MD013 = false # line-length
MD033 = false # inline-html
MD038 = false # no-space-in-code
MD041 = false # first-line-h1
MD046 = false # code-block-style
MD024.allow_different_nesting = true # no-duplicate-headers
MD007.indent = 4 # ul-indent
MD037 = false # no-space-in-emphasis (allow MkDocs Admonitions)

[tool.pymarkdown]
plugins.md013.enabled = false # line-length
plugins.md033.enabled = false # inline-html
plugins.md041.enabled = false # first-line-h1
plugins.md046.enabled = false # code-block-style
plugins.md024.allow_different_nesting = true # no-duplicate-headers
plugins.md007.enabled = true
plugins.md007.indent = 4

[dependency-groups]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio>=1.0.0",
    "pytest-forked>=1.6.0",
    "pytest-timeout==2.3.1",
    "requests==2.32.3",
    "sentence-transformers==3.4.1",
    "aiu-fms-testing-utils>=0.5.0",
    "pytest-mock>=3.15.0",
]
